requirements.txt
flask>=3.0.0
numpy>=1.26.2
opencv-python>=4.8.1.78
pandas>=2.1.3
--find-links https://download.pytorch.org/whl/torch_stable.html
torch>=2.1.0
torchvision>=0.16.0
tensorflow>=2.14.0
Pillow>=10.1.0
streamlit>=1.28.0
requests>=2.31.0 